+++ runner.py
@@ -5,7 +5,7 @@ import log
 from sqlalchemy import create_engine
 from apscheduler.schedulers.blocking import BlockingScheduler
 import pandas as pd
-from scraper import indexScraper, listingScraper
+import scrapers
 
 
 
@@ -15,7 +15,7 @@ def scrape_search_index():
     urls = ['http://www.wg-gesucht.de/en/wohnungen-in-Berlin.8.1.0.0.html',
             'http://www.wg-gesucht.de/en/wohnungen-in-Berlin.8.2.0.0.html']
     for url in urls:
-        listing_html = indexScraper(url).get_html()
+        listing_html = scrapers.indexScraper(url).get_html()
         listings = listing_html.get_listings_from_page()
         n_records = listings.shape[0]
         listings.to_sql('listings_stg', disk_engine, if_exists='append', index=False)
@@ -46,17 +46,28 @@ def write_clean_listings():
     clean_listings.to_sql('listings_clean', disk_engine, if_exists='replace', index=False)
     logging.info('clean listings written')
 
-def run_listing_url_scraper():
+
+def follow_listing_urls():
     """go through listing individual urls to get more detail"""
     # retrieve urls from clean table
     disk_engine = create_engine('sqlite:///database/listings.db')
-    listing_urls = pd.read_sql('select distinct(link) from listings_clean', disk_engine)
-    listing_urls = listing_urls.link.tolist()
-    logging.info('pulled {} distinct urls from local db'.format(len(listing_urls)))
-    # get the further details
+    listings = pd.read_sql('select * from listings_clean', disk_engine)
+    logging.info('pulled {} distinct urls from local db'.format(listings.shape[0]))
+
+    listings.last_scrape_time = pd.to_datetime(listings.last_scrape_time)
+    latest_listing_allowed = listings.last_scrape_time.max() - pd.Timedelta(hours=6)
+    listings = listings.loc[listings.last_scrape_time >= latest_listing_allowed]
+
+    listing_urls = listings.link.tolist()
+    logging.info('list of {} listing URLs to scrape'.format(len(listing_urls)))
+
+    # should functionalise below and go through in batches
+
+    # get further details
     listing_details = []
     for url in listing_urls:
-        listing_details.append(listingScraper(url).get_listing_html().parse_details())
+        scrape_results = scrapers.listingScraper(url).get_listing_html().parse_details()
+        listing_details.append(scrape_results)
         time.sleep(10)
         logging.info('url: {} scraped; sleeping 10 seconds'.format(url))
     listing_details = pd.concat(listing_details)
@@ -68,14 +79,14 @@ def run_scheduler():
     """
     running scheduled scraping and other tasks
     """
-    log.setup_logger('scheduler')
+    log.setup_logger()
 
     scheduler = BlockingScheduler()
-    scheduler.add_job(scrape_search_index, 'interval', minutes=1)
-    scheduler.add_job(write_clean_listings, 'interval', minutes=2)
-
+    #scheduler.add_job(scrape_search_index, 'interval', minutes=2)
+    #scheduler.add_job(write_clean_listings, 'interval', minutes=2)
     # --- currently have individual url scraper disabled - there is a request limit --- #
-    #scheduler.add_job(run_listing_url_scraper, 'interval', minutes=10)
+
+    scheduler.add_job(follow_listing_urls, 'interval', minutes=1)
     print('Press Ctrl+{0} to exit'.format('Break' if os.name == 'nt' else 'C'))
 
     try:
